Split to 3 subcommands, implemented all CLI options for oldWayRelease
